# gen/ledger/gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Ledger API
    description: Record cash inflow and outflow
    version: ""
host: localhost:8080
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
    /openapi.json:
        get:
            tags:
                - ledger
            summary: Download ./gen/http/openapi.json
            operationId: ledger#/openapi.json
            responses:
                "200":
                    description: File downloaded
                    schema:
                        type: file
            schemes:
                - http
    /user/{userID}/expenses:
        post:
            tags:
                - ledger
            summary: RecordExpense ledger
            operationId: ledger#RecordExpense
            parameters:
                - name: userID
                  in: path
                  description: user id
                  required: true
                  type: integer
                  format: int64
                - name: object
                  in: body
                  required: true
                  schema:
                    type: object
                    properties:
                        amount:
                            type: number
                            description: amount spent
                            example: 0.28361020039572515
                            format: double
                        category:
                            type: string
                            description: category
                            example: Enim quasi laudantium autem.
                        memo:
                            type: string
                            description: memo
                            example: Temporibus hic eligendi.
            responses:
                "201":
                    description: Created response.
                "400":
                    description: Bad Request response.
            schemes:
                - http